+++ .env-template
@@ -7,6 +7,7 @@ API_PORT=
 ############################################
 ###             BACKEND uses         	 ###
 ############################################
+API_VERSION=
 MONGO_REF=
 MONGO_PORT=
 MONGO_ROOT_USERNAME=
