+++ app/server/package.json
@@ -19,6 +19,7 @@
     "express": "^4.17.3",
     "express-async-handler": "^1.2.0",
     "express-rate-limit": "^6.4.0",
+    "express-routes-versioning": "^1.0.1",
     "jsonwebtoken": "^8.5.1",
     "mongoose": "^6.2.3",
     "nodemon": "^2.0.15"
